Fix failing test in hicart.n64.ad16
The Pysim-style test cases for this module predate the implementation of
the deep synchronizer for the address-data lines. Additionally, on read,
the high and low words were swapped, as presented to the AD16 bus.

These failures are a good argument for why it must be possible to run
the entire test suite easily and with a single command.